#268 More specimen view work
- Fixed EventOccurrence decimalLatitude & decimalLongitude types
- Updated react-icons to 4.7.1
- Implemented basic specimen details page (WIP)

+++ packages/graphql-api/src/resources/shared/resources/event/event.type.js
@@ -169,8 +169,8 @@ export default gql`
     day: Int
     month: Int
     year: Int
-    decimalLatitude: Int
-    decimalLongitude: Int
+    decimalLatitude: Float
+    decimalLongitude: Float
     institutionCode: String
     locality: String
   }
